Add files via upload
Task: Using JavaScript Constructors, we're going to make multiple simple counters.

Purpose: JavaScript Constructors are a core concept of Web App development. They help make portions of your application 'alive' and can provide instant interaction and updates.

Create an array containing three different strings representing different colors.
Using JavaScript Constructors, append a div into the body of the HTML, and have it so when I click on that div, it increases the number value inside of it by 1, with its starting value at 0 printed inside of it. You will need to use prototypes to make your clicking functionality work.
The div you append to the body should be styled in JS. It should be 100px tall and 200px wide. It should have a border-radius of 10px.
The text inside of the div you create should have a font-size of 18px and should be centered vertically and horizontally in the div.
In your constructor, you should pass a parameter that is the background color of the div.
Loop through your created array and pass through the color to the JavaScript Constructor to give your div that unique background-color.
Submission: Submit the hosted project link to canvas